Done tasks

+++ src/Module1/ObsoleteMeasures.java
@@ -9,8 +9,8 @@ public class ObsoleteMeasures {
         System.out.println("Enter a weight in grams (g):");
         int grams = Integer.parseInt(scanner.nextLine());
 
-        double leiviska = (grams / (20 * 32 * 13.28));
-        double naula = ((grams % (20 * 32 * 13.28)) / (32 * 13.28));
+        int leiviska = (int) (grams / (20 * 32 * 13.28));
+        int naula = (int) ((grams % (20 * 32 * 13.28)) / (32 * 13.28));
         double luoti = ((grams % (32 * 13.28)) / 13.28);
 
         System.out.print(grams + "grams is " + leiviska + "leiviskä, " + naula + "naula and " + luoti + "luoti.");
